+++ src/Assets/SimpleCodeGenerator/Editor/Template.cs
@@ -1,4 +1,7 @@
-﻿using System.IO;
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using JetBrains.Annotations;
 using SimpleCodeGenerator.Core;
 using UnityEditor;
 using UnityEngine;
@@ -8,20 +11,36 @@ namespace SimpleCodeGenerator.Editor
     public class Template
     {
         private readonly string[] content;
+        private readonly Dictionary<string, Template> importedTemplates = new();
 
         private Template(string[] content)
         {
             this.content = content;
         }
 
+        [PublicAPI]
         public static Template ParseFromFile(string path)
         {
             return new Template(File.ReadAllLines(path));
         }
 
+        [PublicAPI]
+        public static Template ParseFromLines(string[] lines)
+        {
+            return new Template(lines);
+        }
+
+        [PublicAPI]
+        public static Template Parse(string contentWithLineBreaks)
+        {
+            return ParseFromLines(contentWithLineBreaks.Split(new[] { "\r\n", "\r", "\n" }, StringSplitOptions.None));
+        }
+
+        [PublicAPI]
         public string Render(object data)
         {
             return new IntermediateTemplate(content, data)
+                .RenderImports()
                 .RenderStandaloneValues()
                 .RenderForLoops()
                 .Build();
@@ -32,6 +51,7 @@ namespace SimpleCodeGenerator.Editor
             return ParseFromFile(GetAbsolutePathToBuiltInTemplate(templateName));
         }
 
+        [PublicAPI]
         public static bool TryFindTemplateInAssets(string templateAssetPath, out Template template)
         {
             if (!templateAssetPath.StartsWith("Assets/"))
@@ -58,6 +78,13 @@ namespace SimpleCodeGenerator.Editor
         internal static AbsolutePath GetAbsolutePathToBuiltInTemplate(string templateName)
         {
             return GetPathToBuiltInTemplates() / $"{templateName}.txt";
+        }
+
+        [PublicAPI]
+        public Template Import(string key, Template template)
+        {
+            importedTemplates.Add(key, template);
+            return this;
         }
     }
 }